vue auth and crud for villages

+++ app/Http/Controllers/Api/VillageController.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\Villages\StoreRequest;
 use App\Http\Requests\Villages\UpdateRequest;
 use App\Models\Photo;
 use App\Models\Presentation;
 use App\Models\Village;
-use Illuminate\Contracts\View\View;
-use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Storage;
@@ -19,43 +18,18 @@ class VillageController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Contracts\View\View
+     * @return \Illuminate\Database\Eloquent\Builder[]|\Illuminate\Database\Eloquent\Collection
      */
-    public function index(Request $request)
+    public function index()
     {
-        $villages = Village::query();
-        $sort_param = $request->get('sort_param');
-        $sort_direction = $request->get('sort_direction');
-        if ($sort_param) {
-            $request->get('order');
-            $villages = $villages->orderBy($sort_param, $sort_direction);
-        }
-        $query = $request->get('q');
-        if ($query) {
-            $villages = $villages->where('name', 'like', '%'.$query.'%');
-        }
-
-        $villages = $villages->paginate(5);
-
-        return view('villages.index', compact('villages'))
-            ->with('i', (request()->input('page', 1) - 1) * 5);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Contracts\View\View
-     */
-    public function create()
-    {
-        return view('villages.create');
+        return Village::with('photo')->get();
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param StoreRequest $request
-     * @return RedirectResponse
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(StoreRequest $request)
     {
@@ -97,41 +71,41 @@ class VillageController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param int $village_id
-     * @return \Illuminate\Contracts\View\View
+     * @param  int  $id
+     * @return \Illuminate\Database\Eloquent\Collection|\Illuminate\Database\Eloquent\Model|null
      */
-    public function show(int $village_id)
+    public function show(int $id)
     {
-        $village = Village::findOrFail($village_id);
-        return view('villages.show', compact('village'));
+        return Village::with('photo')->find($id);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param int $village_id
-     * @return View
+     * @param  int  $id
+     * @return \Illuminate\Database\Eloquent\Collection|\Illuminate\Database\Eloquent\Model|null
      */
-    public function edit(int $village_id)
+    public function edit(int $id)
     {
-        $village = Village::findOrFail($village_id);
-        return view('villages.edit', compact('village'));
+        return Village::with('photo')->find($id);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param UpdateRequest $request
-     * @param int $village_id
-     * @return \Illuminate\Http\RedirectResponse
+     * @param int $id
+     * @return void
      */
-    public function update(UpdateRequest $request, int $village_id)
+    public function update(UpdateRequest $request, int $id)
     {
-        $village = Village::findOrFail($village_id);
+        $village = Village::findOrFail($id);
         $village->update($request->validated());
         $request_photo = $request->file('photo');
         if ($request_photo) {
-            $village->photo->delete();
+            if ($village->photo) {
+                $village->photo->delete();
+            }
             $original_file_name = $request_photo->getClientOriginalName();
             $filename = Str::slug(pathinfo($original_file_name, PATHINFO_FILENAME));
             $ext = $request_photo->getClientOriginalExtension();
@@ -147,9 +121,12 @@ class VillageController extends Controller
 
         $request_presentation = $request->file('presentation');
         if ($request_presentation) {
+            if ($village->presentation) {
+                $village->presentation->delete();
+            }
             $original_file_name = $request_presentation->getClientOriginalName();
             $filename = Str::slug(pathinfo($original_file_name, PATHINFO_FILENAME));
-            $ext = $request_photo->getClientOriginalExtension();
+            $ext = $request_presentation->getClientOriginalExtension();
             $presentation = Presentation::create([
                 'path' => "storage/villages/$village->id/$filename.$ext"
             ]);
@@ -161,29 +138,22 @@ class VillageController extends Controller
         }
 
         $village->save();
-
-
-        return redirect()->route('villages.index')
-            ->with('success', 'Посёлок успешно обновлён.');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param int $village_id
-     * @return RedirectResponse
+     * @param  int  $id
+     * @return \Illuminate\Http\JsonResponse|void
      */
-    public function destroy(int $village_id)
+    public function destroy($id)
     {
-        Storage::deleteDirectory("public/villages/$village_id");
-        $village = Village::findOrFail($village_id);
+        Storage::deleteDirectory("public/villages/$id");
+        $village = Village::findOrFail($id);
         if (count($village->houses)) {
-            return back()->with(['error' => 'К данному посёлку привязаны дома. Пожалуйста, сначала удалите их.']);
+            return response()->json(['error' => 'К данному посёлку привязаны дома. Пожалуйста, сначала удалите их.']);
         }
 
         $village->delete();
-
-        return redirect()->route('villages.index')
-            ->with('success', 'Посёлок успешно удалён.');
     }
 }
